Reverted to v1's definition of the NeighborList.
A vector of pairs is just more convenient for downstream usage than a
pair of vectors. It's easier to iterate over them in a range loop (and
probably more efficient as it's in a single memory allocation), and it's
easier to handle reordering. So we do an "unbreaking" change to get us
back to version 1's definition... sorry.