Updated to subpar v0.3.0 with the renamed parallelize_range().
Also switch back to using a vanilla function instead of a default-defined
macro; the latter is too fragile when the lambda contains complex expressions.

This also means that, to my great regret, we need to restore one of the tests
that I previously removed, to check that the custom macro works correctly.