Check for consistent tie breaking in all exact methods.
This needs a small fix to the VP tree to properly handle tied distances.

We also remove the redundant check of whether the distance is less than
or equal to the limit() in NeighborQueue::add(). It was only there for
efficiency, but if efficiency is a concern, then the caller should
already check this information before even trying to call add().