Document that k= should be sensibly capped in Searcher::search().
This means that implementations aren't obliged to handle impossible
choices of k. It also eliminates the vague wording about the length of
the output indices/distances, as now the output MUST be of length k.

To make it easier for users to comply with the new constraint, we
introduce the cap_k() function to cap the choice of k. We call this
function automatically in find_nearest_neighbors() for safety.